# Introduce cache of suppliers and remove usage of Objects.hash to speed up