--- a/src/main/java/net/datafaker/service/RegExpContext.java
+++ b/src/main/java/net/datafaker/service/RegExpContext.java
@@ -41,12 +41,21 @@ public class RegExpContext {
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
+
         RegExpContext that = (RegExpContext) o;
-        return Objects.equals(exp, that.exp) && Objects.equals(current, that.current) && Objects.equals(root, that.root) && Objects.equals(context, that.context);
+
+        if (!Objects.equals(exp, that.exp)) return false;
+        if (!Objects.equals(current, that.current)) return false;
+        if (!Objects.equals(root, that.root)) return false;
+        return Objects.equals(context, that.context);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(exp, current, root, context);
+        int result = exp != null ? exp.hashCode() : 0;
+        result = 31 * result + (current != null ? current.hashCode() : 0);
+        result = 31 * result + (root != null ? root.hashCode() : 0);
+        result = 31 * result + (context != null ? context.hashCode() : 0);
+        return result;
     }
 }