The problem with `Objects.hash` is that it generates an array each time it is called.

Suppliers could be cached to avoid parsing of expressions each time

it gives speed up for a wide range of methods up to 50 %

expressions
```
before
Benchmark                        Mode  Cnt     Score     Error   Units
DatafakerExpressions.bothify    thrpt   10  5089.609 ± 112.084  ops/ms
DatafakerExpressions.letterify  thrpt   10  5706.392 ± 326.134  ops/ms
DatafakerExpressions.numerify   thrpt   10  6027.549 ±  87.064  ops/ms
DatafakerExpressions.regexify   thrpt   10  2826.806 ±  31.933  ops/ms

VS

after
Benchmark                        Mode  Cnt      Score     Error   Units
DatafakerExpressions.bothify    thrpt   10  10936.056 ±  39.143  ops/ms
DatafakerExpressions.letterify  thrpt   10  13036.411 ±  94.658  ops/ms
DatafakerExpressions.numerify   thrpt   10  13084.755 ± 423.533  ops/ms
DatafakerExpressions.regexify   thrpt   10   3541.279 ±   6.988  ops/ms
```

csv generation
```
before
Benchmark                  Mode  Cnt    Score   Error   Units
DatafakerCsvjson.csvJson  thrpt   10  301.604 ± 2.242  ops/ms

VS

after
Benchmark                  Mode  Cnt    Score    Error   Units
DatafakerCsvjson.csvJson  thrpt   10  377.927 ± 10.929  ops/ms
```


kotlin faker benchmark

```
before
Benchmark                                       Mode  Cnt    Score   Error  Units
Datafaker_KotlinFakerBenchmark.kotlinFakerTest  avgt   10  378.447 ± 1.685  ms/op

VS

after
Benchmark                                       Mode  Cnt    Score   Error  Units
Datafaker_KotlinFakerBenchmark.kotlinFakerTest  avgt   10  324.256 ± 2.438  ms/op
```

simple methods calls
```
before
Benchmark                              Mode  Cnt      Score      Error   Units
DatafakerSimpleMethods.firstname      thrpt   10   5838.567 ±  508.992  ops/ms
DatafakerSimpleMethods.fullname       thrpt   10   2429.474 ±  211.390  ops/ms
DatafakerSimpleMethods.numberBetween  thrpt   10  66001.570 ± 2144.033  ops/ms
DatafakerSimpleMethods.streetAddress  thrpt   10   1724.282 ±   62.587  ops/ms

VS

after
Benchmark                              Mode  Cnt       Score     Error   Units
DatafakerSimpleMethods.firstname      thrpt   10    8972.172 ±  99.469  ops/ms
DatafakerSimpleMethods.fullname       thrpt   10    3232.930 ± 203.367  ops/ms
DatafakerSimpleMethods.numberBetween  thrpt   10  102087.734 ± 655.514  ops/ms
DatafakerSimpleMethods.streetAddress  thrpt   10    2503.758 ±  19.876  ops/ms
```